Convert to Vue single file components
