+++ src/panel/index.js
@@ -1,0 +1,8 @@
+import Terminal from "./components/Terminal.vue";
+import "./index.css";
+
+window.panel.plugin("lukaskleinschmidt/terminal", {
+  sections: {
+    terminal: Terminal,
+  },
+});
